pyproject.toml
[tool.poetry]
name = "pure-app"
version = "0.1.0"
description = ""
authors = ["Maciej Jaworski <maciej@jaworski.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg = {extras = ["c", "pool"], version = "^3.2.1"}
python-decouple = "^3.8"
wait-for-it = "^2.2.2"
whitenoise = "^6.7.0"
django-adminplus = "^0.6"
requests = "^2.32.3"
django-bootstrap5 = "^24.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-django = "^4.8.0"
model-bakery = "^1.18.2"
requests-mock = "^1.12.1"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
